Revert "Fix: fixed configuration issues"
