+++ .ebextensions/python.config
@@ -2,4 +2,4 @@ option_settings:
   aws:elasticbeanstalk:application:environment:
     PYTHONPATH: "/var/app/current:$PYTHONPATH"
   aws:elasticbeanstalk:container:python:
-    WSGIPath: application:app
+    WSGIPath: app:app
